requirements.txt
deep_translator==1.11.4
langchain==0.3.19
langchain_core==0.3.37
langchain_experimental==0.3.4
langchain_openai==0.3.6
matplotlib==3.10.0
numpy==2.2.3
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
pytz==2024.2
seaborn==0.13.2
streamlit==1.41.1
tabulate==0.9.0
openpyxl==3.1.5